feat(css_parser): CSS Parser pseudo element selector #268 - move pseudo class to a module
